% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_B.R
\name{error_B}
\alias{error_B}
\title{Calculating several error measurement of B matrix}
\usage{
error_B(B, B_orig)
}
\arguments{
\item{B}{the estimated B}

\item{B_orig}{ground truth of B}
}
\value{
a named vector of the 11 measurements
}
\description{
There are 11 measurement in total to be measured.
}
\details{
\itemize{
   \item{TP}{:true positive rate}
   \item{TN}{:true negative rate}
   \item{FP}{:false positive rate}
   \item{FN}{:false negative rate}
   \item{SEN}{:sensitivity}
   \item{SPE}{:specificity}
   \item{PREC}{:precision}
   \item{ACC}{:accuracy }
   \item{F1}{:2 * TP/(2*TP + FP + FN)}
   \item{MCC}{:Matthew's Correlation Coefficie}
   \item{MSE}{:mean squared error}
}
}
