% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data.R
\name{g_modelAR1}
\alias{g_modelAR1}
\title{Covariace and precision matrix of AR(1) model
generating precision matrix and covariance of an AR(1) model}
\usage{
g_modelAR1(q, rho = 0.7, tol = 1e-10)
}
\arguments{
\item{q}{number of nodes}

\item{rho}{the covariance of neighborhood}

\item{tol}{tolerence of 0 in precision}
}
\value{
a list with Omega (precision) and Sigma (covariance)
}
\description{
Covariace and precision matrix of AR(1) model
generating precision matrix and covariance of an AR(1) model
}
