% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSSL.R
\name{gSSL}
\alias{gSSL}
\title{Graphical Spike-and-Slab LASSO}
\usage{
gSSL(
  Y,
  xi1 = 0.01 * nrow(Y),
  xi0 = seq(0.1 * nrow(Y), nrow(Y), length = 10),
  eta_hyper_params = c(1, ncol(Y)),
  diag_penalty = FALSE,
  max_iter = 500,
  eps = 0.001,
  obj_counter_max = 5,
  verbose = FALSE
)
}
\arguments{
\item{Y}{response matrix, no design in this case}

\item{xi1}{hyperparameters to be explored by the algorithm, penalty on Omega}

\item{xi0}{a scalar for the slab penalty on Omega}

\item{eta_hyper_params}{hyperparameter to be set, prior on spike weight of Omega}

\item{diag_penalty}{bool, whether to penalize the diagonal, default no}

\item{max_iter}{maximum iterations for the EM algorithm}

\item{eps}{tolerance for convergence}

\item{obj_counter_max}{, maximum number of couting the objective function}

\item{verbose}{bool, whether to print intermidate notes}
}
\value{
A list with dynamic exploration result, point estimates are in `$Omega` and `$B`.
}
\description{
Main posterior exploration algorithm for graphical spike-and-slab LASSO, useful in general chain graph setting for the first layer
}
