% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starSSL.R
\name{idfloor0link}
\alias{idfloor0link}
\title{identity floor 0 truncation link
calculate the latent normal's bound under identity floor 0 truncation link ransformation, i.e. y=floor(z*), y=y*(y>=0), z*~N}
\usage{
idfloor0link(Y)
}
\arguments{
\item{Y}{the observed data}
}
\value{
a list of lower and upper bound of the latent normal
}
\description{
identity floor 0 truncation link
calculate the latent normal's bound under identity floor 0 truncation link ransformation, i.e. y=floor(z*), y=y*(y>=0), z*~N
}
