% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starSSL.R
\name{idfloorlink}
\alias{idfloorlink}
\title{identity floor link
calculate the latent normal's bound under identity floor link ransformation, i.e. y=floor(z*), y=y*(y>=0), z*~N}
\usage{
idfloorlink(Y, upper = TRUE)
}
\arguments{
\item{Y}{the observed data}

\item{upper}{whether to calculate upper bound}
}
\value{
a matrix of lower or upper bound of the latent normal
}
\description{
identity floor link
calculate the latent normal's bound under identity floor link ransformation, i.e. y=floor(z*), y=y*(y>=0), z*~N
}
