% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starSSL.R
\name{logfloorlink}
\alias{logfloorlink}
\title{the logfloor transformation to integer values
calculate the latent normal's bound under log-floor ransformation, i.e. y=floor(exp(z*)), z*~N}
\usage{
logfloorlink(Y)
}
\arguments{
\item{Y}{the observed data}
}
\value{
a list with both lower or upper bound of the latent normal
}
\description{
the logfloor transformation to integer values
calculate the latent normal's bound under log-floor ransformation, i.e. y=floor(exp(z*)), z*~N
}
