% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data.R
\name{mprobit}
\alias{mprobit}
\title{Multivariate probit model
simulating from a multivariate probit model}
\usage{
mprobit(X, B, mu, Sigma, Omega = solve(Sigma))
}
\arguments{
\item{X}{design matrix}

\item{B}{regression coefficients}

\item{mu}{mean vector}

\item{Sigma}{covariance}

\item{Omega}{precision matrix}
}
\value{
a matrix of binary variables sampled from the multivariate probit with mean-covariance parameterization
}
\description{
Multivariate probit model
simulating from a multivariate probit model
}
