% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_support.R
\name{plot_support}
\alias{plot_support}
\title{Stability plot of a dpe result}
\usage{
plot_support(fit_cgSSL_dpe)
}
\arguments{
\item{fit_cgSSL_dpe}{output object from cgSSL that uses condexp=FALSE}
}
\description{
This only makes sense for objects resulting from dynamic posterior exploration. Identify the points that have same support as the terminal point.
}
\details{
Loop over the results in B_path and Omega_path
Identify the points that have same support as the terminal point. If it does, color the point blue, give it pch = 3 (+). If it does not, color the point black, give it pch = 16 (filled circle). If it is unstable, color the point red, give it pch = 4 (x)
}
