% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data.R
\name{rCARm}
\alias{rCARm}
\title{Sampling from CAR, simplest Gaussian chain graph model}
\usage{
rCARm(n, B, X, Sigma)
}
\arguments{
\item{n}{number of samples}

\item{B}{regression coefficients}

\item{X}{Design matrix, should have more than n rows}

\item{Sigma}{Covariance matrix}
}
\value{
a matrix of n rows
}
\description{
Take random samples following CAR model, i.e. N(XBSigma,Sigma)
}
