% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data.R
\name{rcgSTAR}
\alias{rcgSTAR}
\title{simulating a STAR model with latent normal being chain graph parameterization}
\usage{
rcgSTAR(X, B, mu, Sigma, link = function(x) {
    floor(exp(x))
})
}
\arguments{
\item{X}{design matrix}

\item{B}{conditional regression coefficients}

\item{mu}{conditional mean vector}

\item{Sigma}{covariance}

\item{link}{a function convert latent normal to integers, default is the log-floor link}
}
\value{
a matrix of STAR sample
}
\description{
Take random samples following STAR and chain graph parameterization of latent normal, using link function given
}
