% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data.R
\name{rSTAR}
\alias{rSTAR}
\title{simulating a STAR model with latent normal being mean-covariance parameterization}
\usage{
rSTAR(X, B, mu, Sigma, link = function(x) {
    floor(exp(x))
})
}
\arguments{
\item{X}{design matrix}

\item{B}{regression coefficients}

\item{mu}{mean vector}

\item{Sigma}{covariance}

\item{link}{a function convert latent normal to integers, default is the log-floor link}
}
\value{
a matrix of STAR sample
}
\description{
Take random samples following STAR and mean-covariance parameterization of latent normal, using link function given
}
