% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starSSL.R
\name{sqrtfloor0link}
\alias{sqrtfloor0link}
\title{square root floor link truncated at 0
calculate the lower and upper bound when using a square root floor link}
\usage{
sqrtfloor0link(Y)
}
\arguments{
\item{Y}{the data}
}
\value{
a list with lower and upper bound and inverse transformation function
}
\description{
square root floor link truncated at 0
calculate the lower and upper bound when using a square root floor link
}
