% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unitdiag}
\alias{unitdiag}
\title{scale covariance and precision in place
This routine will scale the covariance matrix to have unit diagonal in place and also properly scale the precision.}
\usage{
unitdiag(Sigma, Omega)
}
\arguments{
\item{Sigma}{a matrix that is the covariance matrix, will be scaled to have unit diagonal}

\item{Omega}{should be the inverse of Sigma}
}
\value{
no return, scaling will be done in place
}
\description{
scale covariance and precision in place
This routine will scale the covariance matrix to have unit diagonal in place and also properly scale the precision.
}
