% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_Omega.R
\name{error_Omega}
\alias{error_Omega}
\title{Calculating several error measurement of Omega matrix}
\usage{
error_Omega(Omega, Omega_orig)
}
\arguments{
\item{Omega}{the estimated Omega}

\item{Omega_orig}{ground truth of Omega}
}
\value{
a named vector of the 11 measurements
}
\description{
There are 11 measurement in total to be measured.
}
\details{
\itemize{
 \item{TP}{:true positive rate}
   \item{TN}{:true negative rate}
   \item{FP}{:false positive rate}
   \item{FN}{:false negative rate}
   \item{SEN}{:sensitivity}
   \item{SPE}{:specificity}
   \item{PREC}{:precision}
   \item{ACC}{:accuracy }
   \item{F1}{:2 * TP/(2*TP + FP + FN)}
   \item{MCC}{:Matthew's Correlation Coefficie}
 \item{FROB:} {squared Frobenius error}
}
}
